Separate out CopyPopover & Style it
