+++ src/styles/tools.scss
@@ -35,10 +35,11 @@ $fa-font-path: "~@sparkpost/styles/src/assets/fonts";
 
 .actionLink {
   display: inline-block;
-  padding: 0 rem(9);
+  padding: 0 rem(12);
 
   color: color(orange);
   font-weight: 400;
+  font-size: rem(12);
 
   &:hover {
     cursor: pointer;
